requirements/prod.txt
Django==4.2
django-cleanup==9.0.0
django-notifications-hq==1.8.3
django-star-ratings==0.9.2
dump-env==1.5.0
pillow==11.0.0
python-dotenv==1.0.1
sorl-thumbnail==12.11.0
django_tinymce==4.1.0